Add centralized test data constants and refactor email utility
